Fix staging build (#3398)
* Fix wrong path

* Improve on EMPTY_ENVIRONMENT ambiguity

* Add missing variable

* update base url

* Remove harcoded url

* We don't have enough words in the memory mongo server

* Fix test

* skip e2e integrity tests
